alphabets extracted to separated .hh and .cc files
