+++ src/strings/tests-nfa-segmentation.cc
@@ -27,6 +27,7 @@ using namespace Mata::Nfa;
 using namespace Mata::Strings;
 using namespace Mata::util;
 using namespace Mata::Parser;
+using Symbol = Mata::Symbol;
 
 // Some common automata {{{
